Local jekyll development +  make slightly more mobile friendly
FUTURE NOTE: I don't know what I did to make the .gitignore finally ignore
the _site folder for jekyll, but I think there must have been something wrong
with  it. I told it to ignore all the files individually in SourceTree, and it
worked, so then I opened .gitignore in NotePad++ and removed all the individual
files and just left _site/ and it started ignoring it fine. I'm guessing there
was something wrong with the encoding of the .gitignore file, and it would have
been fixed if I had just deleted the file and recreated it. Correction: I just
checked the old .gitignore and can confirm that it was in a weird encoding.